# docs/openapi-dify.yaml
openapi: 3.0.3
info:
  title: Outdoor SaaS API
  version: 1.0.0
  description: |
    户外广告管理系统 REST API
    用于 Dify AI 助手集成
servers:
  - url: http://121.40.166.153:16000
    description: 本地开发环境（frp内网穿透）
security:
  - authorization: [Bearer eyJhbGciOiJIUzM4NCJ9.eyJzdWIiOiJ0b3VodXdhaSIsInJvbGUiOiJVU0VSIiwiaWF0IjoxNzczMTE0ODE4LCJleHAiOjE3NzMyMDEyMTh9.e_qpuguVudlfaUY-KF1oHflUusMaEwf1kZN3_spuksxK3PebAI511HGOHBel83fn]
paths:
...
  /api/community/{id}:
    get:
      tags:
        - 社区管理
      summary: 根据ID查询社区
      description: 获取指定ID的社区详细信息
      operationId: get_community
      parameters:
        - name: id
          in: path
          required: true
          description: 社区ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCommunity'
        '404':
          description: 社区不存在

    delete:
      tags:
        - 社区管理
      summary: 删除社区
      description: 根据ID删除指定社区
      operationId: delete_community
      parameters:
        - name: id
          in: path
          required: true
          description: 社区ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/community:
    post:
      tags:
        - 社区管理
      summary: 创建社区
      description: 创建新社区
      operationId: create_community
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Community'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'

    put:
      tags:
        - 社区管理
      summary: 更新社区
      description: 更新社区信息
      operationId: update_community
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Community'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/community/filter/page:
    post:
      tags:
        - 社区管理
      summary: 分页查询社区
      description: 根据条件分页查询社区列表
      operationId: filter_communities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageResultCommunity'

  # ========== 道闸管理 ==========
...
  /api/barrier-gate/community/{communityId}:
    get:
      tags:
        - 道闸管理
      summary: 根据社区查询道闸
      description: 获取指定社区下的所有道闸
      operationId: get_barrier_gates_by_community
      parameters:
        - name: communityId
          in: path
          required: true
          description: 社区ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功

  # ========== 框架管理 ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 请输入 JWT Token，格式：Bearer {token}
  schemas:
...
    ResultInteger:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: integer
        timestamp:
          type: integer
          format: int64

    # ========== 社区 ==========
    Community:
      type: object
      properties:
        id:
          type: integer
        communityNo:
          type: string
          description: 社区编号
        buildingName:
          type: string
          description: 楼栋名称
        buildingAddress:
          type: string
          description: 楼栋地址
        city:
          type: string
          description: 城市
        district:
          type: string
          description: 区县
        street:
          type: string
          description: 街道
    ResultCommunity:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/Community'
        timestamp:
          type: integer
          format: int64
...
    CommunityQueryParam:
      type: object
      properties:
        keyword:
          type: string
          description: 关键词搜索
        city:
          type: string
          description: 城市筛选
        pageNum:
          type: integer
          default: 1
        pageSize:
          type: integer
          default: 10

    # ========== 道闸 ==========
...
    PageResult:
      type: object
      properties:
        pageNum:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        pages:
          type: integer
        list:
          type: array
          items:
            type: object
    ResultPageResultCommunity:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          allOf:
            - $ref: '#/components/schemas/PageResult'
            - properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/Community'
        timestamp:
          type: integer
          format: int64

    # ========== 方案关联 ==========
...